♻️ refactor: split cron/cron.go into service.go and persistence.go
- service.go: Service struct, scheduling, job CRUD
- persistence.go: JSON file I/O (loadJobs, saveJobsLocked)
- Delete monolithic cron.go
